[Feature] : log 수정과 refresh token 코드 수정 (#45)

+++ src/main/java/com/potato/balbambalbam/user/join/controller/JoinController.java
@@ -8,6 +8,7 @@ import com.potato.balbambalbam.user.join.jwt.JWTUtil;
 import com.potato.balbambalbam.user.join.service.JoinService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.*;
 
 @Controller
 @ResponseBody
+@Slf4j
 public class JoinController {
 
     private final JoinService joinService;
@@ -34,43 +36,51 @@ public class JoinController {
         return joinService.findUserBySocialId(socialId).getId();
     }
 
+    private String extractSocialIdFromToken(String access) {
+        String socialId = jwtUtil.getSocialId(access);
+        return socialId;
+    }
+
     // 회원정보 받기
     @PostMapping("/users")
     public ResponseEntity<?> createUser(@Validated @RequestBody JoinDto joinDto, HttpServletResponse response) {
 
-        String access = joinService.joinProcess(joinDto, response);
-        response.setHeader("access", access);
+        joinService.joinProcess(joinDto, response); //access, refresh 토큰 생성
+        log.info("회원가입이 완료되었습니다.");
 
-        System.out.println("회원가입이 완료되었습니다.");
         return ResponseEntity.ok().body("회원가입이 완료되었습니다."); //200
     }
 
     // 회원정보 수정
     @PatchMapping("/users")
-    public ResponseEntity<?> updateUser(@Validated @RequestHeader("access") String access, @RequestBody JoinDto joinDto) {
+    public ResponseEntity<?> updateUser(@Validated @RequestHeader("access") String access,
+                                        @RequestBody JoinDto joinDto) {
 
         Long userId = extractUserIdFromToken(access);
         EditDto editUser = joinService.updateUser(userId, joinDto);
+        log.info("{} : 회원정보 수정이 완료되었습니다.", userId);
 
-        System.out.println( userId + " : 사용자 정보가 수정되었습니다.");
         return ResponseEntity.ok().body(editUser); //200
     }
 
     //회원정보 삭제
     @DeleteMapping("/users")
-    public ResponseEntity<?> deleteUser(@RequestHeader("access") String access, @RequestBody DeleteUserDto deleteUserDto,
-                                        HttpServletRequest request, HttpServletResponse response) {
+    public ResponseEntity<?> deleteUser(@RequestHeader("access") String access,
+                                        @RequestBody DeleteUserDto deleteUserDto) {
 
         Long userId = extractUserIdFromToken(access);
         String name = deleteUserDto.getName();
         joinService.deleteUser(userId, name);
 
         // refresh
-        /*if (refresh != null && refreshRepository.existsByRefresh(refresh)) {
+        /*String socialID = extractSocialIdFromToken(access);
+        String refresh = refreshRepository.findRefreshBySocialId(socialID);
+
+        if (refresh != null && refreshRepository.existsByRefresh(refresh)) {
             refreshRepository.deleteByRefresh(refresh);
         }*/
 
-        System.out.println( userId + " : 사용자 정보가 삭제되었습니다.");
+        log.info("{} : 회원 탈퇴가 완료되었습니다.", userId);
         return ResponseEntity.ok().body("회원 탈퇴가 완료되었습니다."); //200
     }
 
@@ -81,7 +91,7 @@ public class JoinController {
         Long userId = extractUserIdFromToken(access);
         EditDto editUser = joinService.findUserById(userId);
 
-        System.out.println( userId + " : 사용자 정보를 전송했습니다.");
+        log.info("{} : 회원정보가 출력되었습니다.", userId);
         return ResponseEntity.ok().body(editUser); //200
     }
