+++ src/main/java/com/potato/balbambalbam/user/join/service/JoinService.java
@@ -1,7 +1,7 @@
 package com.potato.balbambalbam.user.join.service;
 
+import com.potato.balbambalbam.data.entity.Refresh;
 import com.potato.balbambalbam.data.entity.User;
-import com.potato.balbambalbam.data.entity.WeakSoundTestStatus;
 import com.potato.balbambalbam.data.repository.*;
 import com.potato.balbambalbam.exception.InvalidUserNameException;
 import com.potato.balbambalbam.exception.SocialIdChangeException;
@@ -12,6 +12,8 @@ import com.potato.balbambalbam.user.join.jwt.JWTUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
+
+import java.util.Date;
 
 @Service
 public class JoinService {
@@ -48,19 +50,15 @@ public class JoinService {
     }
 
     //새로운 회원정보 저장
-    public String joinProcess(JoinDto joinDto, HttpServletResponse response) {
+    public void joinProcess(JoinDto joinDto, HttpServletResponse response) {
 
         String name = joinDto.getName();
         String socialId = joinDto.getSocialId();
         Integer age = joinDto.getAge();
         Byte gender = joinDto.getGender();
 
-        Boolean isExist = userRepository.existsBySocialId(socialId);
-        if (isExist) {return null;}
-
-        User data = new User();
-
         //데이터베이스에 회원정보 저장
+        User data = new User();
         data.setName(name);
         data.setSocialId(socialId);
         data.setAge(age);
@@ -69,19 +67,17 @@ public class JoinService {
         userRepository.save(data);
 
         //access 토큰 발급
-        String access = jwtUtil.createJwt("access", socialId, data.getRole(), 7200000L); //120분
+        String access = jwtUtil.createJwt("access", socialId, data.getRole(), 7200000L); // 7200000L 120분
 
         // Refresh 토큰 발급
-        /*String refresh = jwtUtil.createJwt("refresh", socialId, data.getRole(), 86400000L); // 24시간
-        addRefreshEntity(socialId, refresh, 86400000L);*/
+        String refresh = jwtUtil.createJwt("refresh", socialId, data.getRole(), 86400000L); // 86400000L 24시간
+        addRefreshEntity(socialId, refresh, 86400000L);
 
-        /*response.setHeader("access", access);
-        response.setHeader("refresh", refresh);*/
-
-        return access;
+        response.setHeader("access", access);
+        response.setHeader("refresh", refresh);
     }
 
-    /*private void addRefreshEntity(String socialId, String refresh, Long expiredMs) {
+    private void addRefreshEntity(String socialId, String refresh, Long expiredMs) {
 
         Date date = new Date(System.currentTimeMillis() + expiredMs);
 
@@ -91,7 +87,7 @@ public class JoinService {
         refreshEntity.setExpiration(date.toString());
 
         refreshRepository.save(refreshEntity);
-    }*/
+    }
 
     // 회원정보 업데이트
     @Transactional
@@ -124,13 +120,28 @@ public class JoinService {
             throw new InvalidUserNameException("닉네임이 일치하지 않습니다."); //400
         }
 
-        cardBookmarkRepository.deleteByUserId(userId);
-        cardScoreRepository.deleteByUserId(userId);
-        cardWeakSoundRepository.deleteByUserId(userId);
 
-        customCardRepository.deleteUserById(userId);
-        userWeakSoundRepository.deleteByUserId(userId);
-        weakSoundTestSatusRepositoy.deleteByUserId(userId);
+        if(cardBookmarkRepository.existsByUserId(userId)){
+            cardBookmarkRepository.deleteByUserId(userId);
+
+        }
+        if(cardScoreRepository.existsByUserId(userId)){
+            cardScoreRepository.deleteByUserId(userId);
+        }
+        if(cardWeakSoundRepository.existsByUserId(userId)){
+            cardWeakSoundRepository.deleteByUserId(userId);
+
+        }
+        if(customCardRepository.existsByUserId(userId)){
+            customCardRepository.deleteUserById(userId);
+        }
+        if(userWeakSoundRepository.existsByUserId(userId)){
+            userWeakSoundRepository.deleteByUserId(userId);
+
+        }
+        if(weakSoundTestSatusRepositoy.existsByUserId(userId)){
+            weakSoundTestSatusRepositoy.deleteByUserId(userId);
+        }
 
         userRepository.deleteById(userId);
